+++ tests/turbo.node.test.ts
@@ -1,11 +1,17 @@
 import { ArweaveSigner, createData } from 'arbundles';
-import Arweave from 'arweave';
 import { CanceledError } from 'axios';
 import { expect } from 'chai';
 import fs from 'fs';
+import { describe } from 'mocha';
 import { Readable } from 'node:stream';
+import { restore, stub } from 'sinon';
 
 import { USD } from '../src/common/currency.js';
+import {
+  ARToTokenAmount,
+  ArweaveToken,
+  WinstonToTokenAmount,
+} from '../src/common/token.js';
 import {
   TurboAuthenticatedClient,
   TurboUnauthenticatedClient,
@@ -13,13 +19,24 @@ import {
 import { TurboFactory } from '../src/node/factory.js';
 import { FailedRequestError } from '../src/utils/errors.js';
 import {
+  delayedBlockMining,
   expectAsyncErrorThrow,
+  fundArLocalWalletAddress,
+  getRawBalance,
+  mineArLocalBlock,
+  sendFundTransaction,
+  testArweave,
   testJwk,
   testWalletAddress,
   turboDevelopmentConfigurations,
 } from './helpers.js';
 
 describe('Node environment', () => {
+  afterEach(() => {
+    // Restore all stubs
+    restore();
+  });
+
   describe('TurboFactory', () => {
     it('should return a TurboUnauthenticatedClient when running in Node environment and not provided a privateKey', () => {
       const turbo = TurboFactory.unauthenticated(
@@ -220,21 +237,101 @@ describe('Node environment', () => {
         expect(winc).to.be.a('string');
       });
     });
+
+    describe('submitFundTransaction()', () => {
+      before(async () => {
+        await fundArLocalWalletAddress(testWalletAddress);
+
+        await mineArLocalBlock();
+      });
+
+      it('should return a FailedRequestError when submitting a non-existent payment transaction ID', async () => {
+        const nonExistentPaymentTxId = 'non-existent-payment-tx-id';
+        const error = await turbo
+          .submitFundTransaction({ txId: nonExistentPaymentTxId })
+          .catch((error) => error);
+        expect(error).to.be.instanceOf(FailedRequestError);
+        expect(error.message).to.contain('Failed request: 404: Not Found');
+      });
+
+      it('should properly submit an existing payment transaction ID to the Turbo Payment Service for processing a pending tx', async () => {
+        const txId = await sendFundTransaction(1000);
+
+        const { id, winc, owner, token, status } =
+          await turbo.submitFundTransaction({
+            txId,
+          });
+        expect(id).to.equal(txId);
+        expect(owner).to.equal(testWalletAddress);
+        expect(winc).to.equal('766');
+        expect(token).to.equal('arweave');
+        expect(status).to.equal('pending');
+      });
+
+      const minConfirmations = 25;
+      it('should properly submit an existing payment transaction ID to the Turbo Payment Service for processing a confirmed tx', async () => {
+        const balanceBefore = await getRawBalance(testWalletAddress);
+
+        const txId = await sendFundTransaction(1000);
+        await mineArLocalBlock(minConfirmations);
+
+        const { id, winc, owner, token, status } =
+          await turbo.submitFundTransaction({
+            txId,
+          });
+        expect(id).to.equal(txId);
+        expect(owner).to.equal(testWalletAddress);
+        expect(winc).to.equal('766');
+        expect(token).to.equal('arweave');
+        expect(status).to.equal('confirmed');
+
+        const balanceAfter = await getRawBalance(testWalletAddress);
+
+        expect(+balanceAfter - +balanceBefore).to.equal(766);
+      });
+    });
   });
 
   describe('TurboAuthenticatedNodeClient', () => {
     let turbo: TurboAuthenticatedClient;
 
+    const arweaveToken = new ArweaveToken({
+      arweave: testArweave,
+      pollingOptions: {
+        maxAttempts: 3,
+        pollingIntervalMs: 10,
+        initialBackoffMs: 0,
+      },
+    });
+    const tokenMap = {
+      arweave: arweaveToken,
+    };
+
     before(async () => {
       turbo = TurboFactory.authenticated({
         privateKey: testJwk,
         ...turboDevelopmentConfigurations,
+        // @ts-ignore
+        tokenMap,
       });
     });
 
-    it('getBalance()', async () => {
-      const balance = await turbo.getBalance();
-      expect(+balance.winc).to.equal(0);
+    describe('getBalance()', async () => {
+      it('returns correct balance for test wallet', async () => {
+        const rawBalance = await getRawBalance(testWalletAddress);
+        const balance = await turbo.getBalance();
+        expect(balance.winc).to.equal(rawBalance);
+      });
+
+      it('returns correct balance for an empty wallet', async () => {
+        const emptyJwk = await testArweave.crypto.generateJWK();
+        const emptyTurbo = TurboFactory.authenticated({
+          privateKey: emptyJwk,
+          ...turboDevelopmentConfigurations,
+        });
+        const balance = await emptyTurbo.getBalance();
+        expect(balance.winc).to.equal('0');
+      });
     });
 
     describe('uploadFile()', () => {
@@ -250,7 +347,8 @@ describe('Node environment', () => {
           ],
         },
         {
-          target: 'WeirdCharacters-_!felwfleowpfl12345678901234',
+          // cspell:disable
+          target: 'WeirdCharacters-_!felwfleowpfl12345678901234', // cspell:disable
           anchor: 'anchor-MusTBe__-__TwoBytesLong!!',
           tags: [
             {
@@ -375,7 +473,7 @@ describe('Node environment', () => {
       });
 
       it('should return a FailedRequestError when the file is larger than the free limit and wallet is underfunded', async () => {
-        const nonAllowListedJWK = await Arweave.crypto.generateJWK();
+        const nonAllowListedJWK = await testArweave.crypto.generateJWK();
         const filePath = new URL('files/1MB_file', import.meta.url).pathname;
         const fileSize = fs.statSync(filePath).size;
         const newTurbo = TurboFactory.authenticated({
@@ -402,7 +500,9 @@ describe('Node environment', () => {
         .catch((error) => error);
       expect(error).to.be.instanceOf(FailedRequestError);
       // TODO: Could provide better error message to client. We have error messages on response.data
-      expect(error.message).to.equal('Failed request: 400: Bad Request');
+      expect(error.message).to.equal(
+        "Failed request: 400: No promo code found with code 'BAD_PROMO_CODE'",
+      );
     });
 
     it('getWincForFiat() without a promo could return proper rates', async () => {
@@ -423,7 +523,56 @@ describe('Node environment', () => {
           })
           .catch((error) => error);
         expect(error).to.be.instanceOf(FailedRequestError);
-        expect(error.message).to.equal('Failed request: 400: Bad Request');
+        expect(error.message).to.equal(
+          "Failed request: 400: No promo code found with code 'BAD_PROMO_CODE'",
+        );
+      });
+    });
+
+    describe('fund()', function () {
+      it('should succeed to fund account using arweave tokens', async () => {
+        const [{ winc }] = await Promise.all([
+          turbo.topUpWithTokens({
+            tokenAmount: WinstonToTokenAmount(10),
+          }),
+          delayedBlockMining(),
+        ]);
+
+        expect(winc).to.equal('7');
+      });
+
+      it('should fail to submit fund tx when arweave fund tx is stubbed to succeed but wont exist on chain', async () => {
+        stub(arweaveToken, 'submitTx').resolves();
+
+        // simulate polling for transaction
+        stub(testArweave.api, 'post')
+          .onFirstCall()
+          .throws()
+          .onSecondCall()
+          .resolves(undefined)
+          .onThirdCall()
+          .resolves({ data: { data: { transaction: true } } } as any);
+
+        const error = await turbo
+          .topUpWithTokens({
+            tokenAmount: ARToTokenAmount(0.000_000_100_000),
+            feeMultiplier: 1.5,
+          })
+          .catch((error) => error);
+        expect(error).to.be.instanceOf(Error);
+        expect(error.message).to.contain('Failed to submit fund transaction!');
+      });
+
+      it('should fail to submit fund tx when fund tx fails to post to arweave', async () => {
+        stub(testArweave.transactions, 'post').throws();
+
+        const error = await turbo
+          .topUpWithTokens({
+            tokenAmount: WinstonToTokenAmount(1000),
+          })
+          .catch((error) => error);
+        expect(error).to.be.instanceOf(Error);
+        expect(error.message).to.contain('Failed to post transaction');
       });
     });
   });